feat: CollapsibleContentコンポーネントを実装
